Revert adding tests
