--- server/.env.test
@@ -1,6 +1,0 @@
-# These are environment variables for development in local machine.
-# Environment variables for production should be written in app.yaml.
-
-GOOGLE_CLOUD_PROJECT=gpay-group-buy
-SERVER_URL=http://localhost:5000
-WEB_URL=http://localhost:3000
