+++ server/tests/integration/customers.test.ts
@@ -16,7 +16,7 @@
 
 import request from 'supertest';
 
-import app from '../../src/app';
+import app from '../../src';
 import customerFixtures from '../fixtures/customers';
 
 describe('Customers endpoints', () => {
